fix(product): split stick-to-bottom test file under the max-lines cap
use-transcript-stick-to-bottom.test.tsx had grown to 611 lines (cap is
600). Moves the PRO-175 submit-stamp session-scoping describe block,
plus its own small dedicated harness, into a sibling
use-transcript-stick-to-bottom.session-stamp.test.tsx. No allowlist
touched; both files are now well under the cap.
